test(coordinator): follow the work ACROSS a stage boundary — and fix what that found
There was no test that asked "does stage 2 actually start?", and five separate
bugs hid behind its absence. TestIntegration_EndToEnd_SimplePath creates ONE
task, runs a mock provider and completes it — it never crosses a seam, so each
bug stayed invisible until the one in front of it was fixed:

  1. ProcessApprovalRequest never called OnAgentApproved        (fixed 09-07)
  2. the handoff went to the THREAD collection, never polled    (fixed 09-07)
  3. the row was stored but no notification published           (fixed 09-07)
  4. the handoff was deduplicated against its own parent        (fixed 09-14)
  5. the handoff never named the artifact it handed off         (fixed 09-14)

Every one is plumbing between two components that each worked, and not one
needed a model, a network or a cloud plane to reproduce — only a test that
followed the work across the boundary.

pipeline_stage_crossing_test.go does that, with the real SQLite coordinator
store and the real messaging store: approve stage 1, then assert stage 2 has
something to pick up IN THE PLACE DISPATCH LOOKS, that it carries the parent
link dedup needs, that it names the artifact, and that the task built from it
is not suppressed. Plus the partition property — an auto edge must fire at
completion and NOT again at approval.

VERIFIED AS AN INSTRUMENT, by reverting each bug in turn:

  bug 4 -> "the stage-2 task was suppressed by task-08032ebc (completed) —
            the pipeline cannot advance"
  bug 5 -> "handoff does not name the artifact
            design_docs/planned/m-secondary-model-fallback.md"
  bug 2 -> "approving stage 1 put NOTHING in the next stage's inbox"

A test that has never failed is a claim, not a guard.

AND IT FOUND A SIXTH, immediately: SetTaskFingerprint passed a raw uint64 to
go-sqlite3, which refuses values with the high bit set ("uint64 values with
high bit set are not supported"). simhash returns a full 64-bit value, so
roughly HALF of all content failed to store a fingerprint — and both callers
discarded the error, so a task simply had none. A task with no fingerprint can
never be matched, which is indistinguishable from "no duplicate found": dedup
did not apply to half the traffic on any SQLite coordinator, silently, and the
read path had the same defect so a stored value would not have matched anyway.

Fixed by casting on both write and read, which is what Firestore has always
done — making the two stores agree rather than inventing a third behaviour.
Values below 2^63 are unchanged, so existing fingerprints keep matching. The
callers now log instead of discarding, because a task that cannot suppress a
future duplicate is worth one line.

+++ internal/coordinator/store_sqlite.go
@@ -690,7 +690,7 @@ func (s *SQLiteStore) FindDuplicateTask(ctx context.Context, fingerprint uint64,
 		        error, output, cost, tokens_used,
 		        capabilities_json, impact_level, estimated_cost
 		FROM tasks WHERE fingerprint = ? ORDER BY created_at DESC LIMIT ?`,
-		fingerprint, DedupCandidateLimit,
+		int64(fingerprint), DedupCandidateLimit, // int64 on BOTH sides — see SetTaskFingerprint
 	)
 	if err != nil {
 		return nil, err
@@ -713,10 +713,25 @@ func (s *SQLiteStore) FindDuplicateTask(ctx context.Context, fingerprint uint64,
 }
 
 // SetTaskFingerprint sets the fingerprint for duplicate detection
+// SetTaskFingerprint stores the dedup fingerprint.
+//
+// int64, not uint64. go-sqlite3 REFUSES a uint64 with the high bit set —
+// "converting argument $1 type: uint64 values with high bit set are not
+// supported" — and simhash returns a full 64-bit value, so roughly half of all
+// content failed to store a fingerprint at all. Both callers in
+// daemon_tasks_polling discarded the error, so the effect was silent: a task
+// with no stored fingerprint can never be matched, and dedup simply did not
+// apply to half the traffic on any SQLite coordinator.
+//
+// The cast is what Firestore has always done (int64(fingerprint) in
+// coordinator_transitions.go), so this makes the two stores agree rather than
+// inventing a third behaviour. Values below 2^63 are unchanged, so fingerprints
+// already stored keep matching; higher ones wrap to negative consistently on
+// both write and read.
 func (s *SQLiteStore) SetTaskFingerprint(ctx context.Context, id string, fingerprint uint64) error {
 	_, err := s.db.ExecContext(ctx,
 		"UPDATE tasks SET fingerprint = ? WHERE id = ?",
-		fingerprint, id,
+		int64(fingerprint), id,
 	)
 	return err
 }
